Fix integer types to avoid warnings.

+++ lib/fles_ipc/StorableMicroslice.cpp
@@ -21,6 +21,6 @@ StorableMicroslice::StorableMicroslice(MicrosliceDescriptor d,
                                        std::vector<uint8_t> content_v)
     : _desc(d), _content{std::move(content_v)}
 {
-    _desc.size = _content.size();
+    _desc.size = static_cast<uint32_t>(_content.size());
 }
 }
